chore(*): rename exception to exceptions
