+++ src/exception/http/mod.ts
@@ -1,4 +1,4 @@
 // created from ctix
 
 export * from './enum.ts';
-export * from './exception.ts';
+export * from './exceptions.ts';
